Refactor inference pipeline and update model artifacts
- Update `FeatureEngineerMixin` to handle missing columns with type-specific defaults and improve imputation logic.
- Configure scikit-learn to output pandas DataFrames globally in `main.py`.
- Remove manual mixin injection into `__main__` in the application lifespan.
- Update model and pipeline joblib files for the `modelling_phase_v5` ensemble.

+++ server/src/schema/predictions.py
@@ -50,8 +50,8 @@ class PredictionResult(BaseModel):
     respiratory_disease_rate: float
     cardio_mortality_rate: float
     vector_disease_risk_score: float
-    waterborne_disease_incidents: float
-    heat_related_admissions: float
+    waterborne_disease_incidents: int
+    heat_related_admissions: int
 
 
 class SimulationResponse(BaseModel):
